Move section, header, and activity panel out of layout folder
