+++ client/header/activity-panel/README.md
