+++ client/header/activity-panel/activity-card/README.md
