+++ client/header/activity-panel/activity-card/placeholder.js
