+++ client/header/activity-panel/activity-card/style.scss
