+++ client/header/activity-panel/activity-card/test/placeholder.js
