+++ client/header/activity-panel/activity-header/README.md
