+++ client/header/activity-panel/activity-header/style.scss
