+++ client/header/activity-panel/activity-outbound-link/README.md
