+++ client/header/activity-panel/activity-outbound-link/index.js
