+++ client/header/activity-panel/activity-outbound-link/style.scss
