+++ client/header/activity-panel/panels/stock.js
