+++ client/header/activity-panel/style.scss
