+++ client/header/activity-panel/toggle-bubble.js
