+++ client/header/activity-panel/wordpress-notices.js
